requirements.txt
asgiref==3.8.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.0.4
django-allauth==0.62.1
django-cors-headers==4.3.1
django-environ==0.11.2
django-timezone-field==6.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
install==1.3.5
pillow==10.3.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
wcwidth==0.2.13
